+++ client/lib/error-types.js
@@ -1,0 +1,7 @@
+module.exports = {
+    VALIDATION_FAILURE: 'validation_failure',
+    COMMUNICATION_ISSUE: 'communication_issue',
+    INVALID_TOKEN: 'invalid_token',
+    INSUFFICIENT_SCOPE: 'insufficient_scope',
+    INVALID_ARGUMENTS: 'invalid_arguments',
+};
