+++ package.json
@@ -22,6 +22,7 @@
   ],
   "scripts": {
     "coverage": "nyc npm run test",
+    "coverage-html": "nyc --reporter=html npm run test",
     "test": "./node_modules/.bin/mocha --exit --timeout 20000 test/**/*.js",
     "lint": "./node_modules/.bin/jshint index.js",
     "lint-fix": "./node_modules/.bin/js-beautify -r index.js"
